+++ src/lang/hum/gram/mod.rs
@@ -1,0 +1,11 @@
+// devela::lang::hum::grammar
+//
+//! Grammar, structural theory.
+//
+
+// WIPZONE
+crate::structural_mods! { // _mods
+    _mods {
+        // WIPZONE:
+    }
+}