+++ src/lang/hum/nat/en/mod.rs
@@ -1,0 +1,14 @@
+// devela::lang::hum::nat::en
+//
+//! English language.
+//
+
+// WIPZONE
+// mod tense;
+
+crate::structural_mods! { // _mods
+    _mods {
+        // WIPZONE:
+        // pub use super::tense::*;
+    }
+}